+++ db/migrate/20200311100136_add_manual_to_checks.rb
@@ -1,0 +1,5 @@
+class AddManualToChecks < ActiveRecord::Migration[6.0]
+  def change
+    add_column :tasks, :manual, :boolean
+  end
+end